+++ src/coro_lib_test/st_suspend_forever_test.cpp
@@ -12,7 +12,7 @@
 
 namespace
 {
-  coro::co<int> async_simple(coro::st::context & ctx)
+  coro::co<size_t> async_simple(coro::st::context & ctx)
   {
     auto x = co_await coro::st::async_wait_any(ctx,
       coro::st::async_suspend_forever,
